+++ module/utilities/game-settings.js
@@ -43,18 +43,43 @@ export async function registerGameSettings() {
     config: true
   });
 
-  game.settings.register("cyphersystem", "showRuler", {
-    name: game.i18n.localize("CYPHERSYSTEM.SettingShowRuler"),
-    hint: game.i18n.localize("CYPHERSYSTEM.SettingShowRulerHint"),
+  game.settings.register("cyphersystem", "showRulerGridless", {
+    name: game.i18n.localize("CYPHERSYSTEM.SettingShowRulerGridless"),
+    hint: game.i18n.localize("CYPHERSYSTEM.SettingShowRulerGridlessHint"),
     scope: "world",
     type: Number,
     default: 1,
     requiresReload: true,
     choices: {
-      0: game.i18n.localize("CYPHERSYSTEM.SettingShowRulerNever"),
-      1: game.i18n.localize("CYPHERSYSTEM.SettingShowRulerInCombat"),
-      2: game.i18n.localize("CYPHERSYSTEM.SettingShowRulerAlways")
+      0: game.i18n.localize("CYPHERSYSTEM.SettingShowRulerGridlessHide"),
+      1: game.i18n.localize("CYPHERSYSTEM.SettingShowRulerGridlessInCombat"),
+      2: game.i18n.localize("CYPHERSYSTEM.SettingShowRulerGridlessAlways")
     },
+    config: true
+  });
+
+  game.settings.register("cyphersystem", "disableRulerTypes", {
+    name: game.i18n.localize("CYPHERSYSTEM.SettingsDisableRulerTypes"),
+    hint: game.i18n.localize("CYPHERSYSTEM.SettingsDisableRulerTypesHint"),
+    scope: "world",
+    config: true,
+    type: String,
+    default: "marker",
+    requiresReload: true
+  });
+
+  game.settings.register("cyphersystem", "tokenSpeed", {
+    name: game.i18n.localize("CYPHERSYSTEM.SettingTokenSpeed"),
+    hint: game.i18n.localize("CYPHERSYSTEM.SettingTokenSpeedHint"),
+    scope: "world",
+    type: new foundry.data.fields.NumberField({
+      min: 0.5,
+      max: 5,
+      step: 0.5,
+      initial: 1,
+      nullable: false
+    }),
+    requiresReload: true,
     config: true
   });